Default LOGIN_DISABLED to False
Previously this default was derived from the value of TESTING when no
expicit value was found. However this is not desirable when testing a
code base which seeks to make use of the extension. Before
LOGIN_DISABLED had to be manually set to False to avoid this. Now this
is the default.

This may break existing code if the assumption is that TESTING will
determine whether the extension is enabled or not. Please make sure to
take this into consideration prior to upgrading!