Fix handling of X-Forwarded-For header
The code that calculates a user's identifier uses the remote address or
the contents of the X-Forwarded-For header doesn't account for the fact
that the X-Forwarded-For header can contain a comma separated list of
remote addresses, which is the case for scenarios in which 2+ proxy
servers are involved in the request forwarding. This caveat causes
Flask-Login to unexpectedly log users out when the combination of proxy
servers that routed a previous request change.

See http://tools.ietf.org/html/rfc7239 for more information about
X-Forwarded-For standard.