docs: Point Flask Snippet URL to Wayback Machine
Since the docs were written, the Flask project reworked their
infrastructure underneath what is now the Pallets Project. The snippets
part of their servers appear to have been lost and discarded. So, this
commit changes the link in the Flask-Login docs to point to the last
known good copy of the page in the Wayback Machine.

Signed-off-by: Justin W. Flory <git@jwf.io>