Do not pass request arguments to login view except in dynamic subdomain special case
In #462, the requested view arguments were passed to the `login_view` so as to allow for the use of dynamic subdomains in Flask. However, this had the side effect of passing any requested view arguments to the `login_view`. This is unnecessary and may also lead to unanticipated side effects if some arguments inadvertently passed take on special meaning in the `login_view`.

Instead, special case dynamic subdomain view arguments only so that they are the only ones passed to `login_view`.