Added py.typed file for improved type hinting and compatibility
In this commit, a new file named py.typed has been added to the project. This file is utilized to enhance type hinting and compatibility for Python code. By including this file in the project's root directory, we are signaling to tools and libraries that our codebase is fully type-annotated using Python-type hints.

More details: https://peps.python.org/pep-0561/#packaging-type-information